Enables support for distributing Lottie as a binary library
This fixes #1081.

**Solution**

To allow frameworks to be used by multiple Swift version, we need to set:

```
BUILD_LIBRARY_FOR_DISTRIBUTION = YES
```

This isn't necessary for a Swift package (since they are build from source),
but it's useful for Carthage or other means of distributing binary frameworks.

**Testing**

I built `Lottie.framework` in Xcode 11.3.1. Then in Xcode 11.4 I created a
test project and linked to the generated project. It built and linked
successfully.