Fix layout jump
If this is called multiple times, currently, we will get the ending
position as the 'from' value. This results in a potential visual 'jump'
in the view. But by using the presentation layer values, we get the
values for where the layer is currently on screen, regardless of how
many times this ends up running.