Fix frame setting issue for CompatibleAnimationView.
As discussed on the GitHub page in the past, CompatibleAnimationView does not correctly set up the constaints as to make the underlying AnimationView fill the entirity of the frame provided to CompatibleAnimationView.
https://github.com/airbnb/lottie-ios/issues/1091

This fix is the same that was discussed in the issue thread and resolves the frame setting issue.