Core Animation Engine - Set currentFrame when animation is finished
For the Core Animation Engine the `currentFrame` is calculated based
on the `animationProgress` placeholder animation.
The `animationProgress` value is never copied from the presentation layer
to the model layer.

Because of that value of `currentFrame` is lost once the place holder animation
has reached completion and was automatically removed.

This PR changes the placeholder animation to remain and instead be manually
removed in the `animationDidStop(..)` delegate.
That way it is possible to persist the `currentFrame` value.

Attempts to solve issues: